requirements_ocr.txt
pandas>=1.3.0
numpy>=1.21.0
opencv-python>=4.5.0
paddleocr>=2.7.0
paddlepaddle
pytesseract
Pillow
